Added `de/maps` module
